DEP: plat/kvm/arm: Enable the Unikernel to be built as an UEFI application
If `CONFIG_EFI_STUB` is enabled, add to the build system the ARM UEFI
entry stub and the ARM specific post-EFI stub that applies finishing
touches to the platform initialization.

Furthermore, in order for things to work, add the `.data.boot`
sections into the ARM linker script. Similarly to how x86 does, this
is where boot specific stuff will be placed, at the very beginning
of the binary, making this a very convenient place to place the PE
header at.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>