DEP: plat/kvm: Add `QEMU Virt` specific boot configuration entry
Add a configuration entry `KVM_BOOT_QEMU_VIRT` to explicitly
indicate that we are booting from the default `QEMU Virt`
environment.

Furthermore, change the name of the entry point symbol from
`_libkvmplat_entry` to `qemu_virt_entry` to reflect that and remove
the hardcoded linker script ELF entry and rely on the `KVM_LDFLAGS`
variable instead to decide the symbol representing the entry point.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>