DEP: arch/arm/arm64: Fix `stxr`'s used in `spinlock`s register size
Our `spinlock`s have a size of 4-bytes and an alignment of 8, which,
in our `semaphore` implementation, causes a padding of 4 bytes. This
is fine, but it is an inconsistency.

With `QEMU` direct kernel boot, all of our free memory is zeroed out.
However, when using a previous boot phase such as `UEFI`, this
inconsistency shows itself through an always locked `spinlock`.
`UEFI` firmware "poisons" all of its free memory with `0xaf` bytes
which leads to our `stxr`, whose second register argument is 64-bit
instead of 32-bit, also atomically storing the `poisoned` padding.

Thus, make sure that our `stxr` uses 32-bit register for the to be
transferred register.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>
GitHub-Fixes: #289